Avoid spurious sched wakeup IPIs
Because mwait_idle_with_hints() gets called from '!idle context' it must
call current_clr_polling(). This however means that resched_task() is
very likely to send an IPI even when we were polling.
So while it is correct (extra IPIs aren't a problem, whereas a missed
IPI would be) it is a performance problem.
Avoid this issue by using fetch_or() to atomically set NEED_RESCHED
and test if POLLING_NRFLAG is set.
Since a CPU stuck in mwait is unlikely to modify the flags word, contention
on the cmpxchg is unlikely and thus we should mostly succeed in a single go.